src/pipeline/requirements.txt
pyspark==3.5.1
pytest==8.3.2
python-dotenv==1.0.1